+++ src/test/test.snippets.ts
@@ -1,3 +1,4 @@
+import { Compare } from 'klingsten-snippets';
 // import { removeNonPrint } from '../snippets';
 export class TestSnippets {